refactor(examples): move the avatar generator out of the scaffolded config
